move logic to CleanAPI
