+++ cg/cli/clean.py
@@ -2,7 +2,7 @@
 import logging
 from datetime import datetime, timedelta
 from pathlib import Path
-from typing import List, Optional, Iterator
+from typing import List, Optional
 
 import click
 from alchy import Query
@@ -23,11 +23,11 @@ from cg.cli.workflow.commands import (
 )
 from cg.constants import FlowCellStatus, HousekeeperTags
 from cg.constants.constants import DRY_RUN, Sequencers
-from cg.constants.tags import WORKFLOW_PROTECTED_TAGS
+from cg.meta.clean.api import CleanAPI
 from cg.meta.clean.demultiplexed_flow_cells import DemultiplexedRunsFlowCell
 from cg.meta.clean.flow_cell_run_directories import RunDirFlowCell
 from cg.models.cg_config import CGConfig
-from cg.store import Store, models
+from cg.store import Store
 from housekeeper.store import models as hk_models
 
 CHECK_COLOR = {True: "green", False: "red"}
@@ -114,7 +114,7 @@ def _get_confirm_question(bundle, file_obj):
 @DRY_RUN
 @click.pass_context
 def scout_finished_cases(
-        context: click.Context, days_old: int, yes: bool = False, dry_run: bool = False
+    context: click.Context, days_old: int, yes: bool = False, dry_run: bool = False
 ):
     """Clean up of solved and archived scout cases"""
     scout_api: ScoutAPI = context.obj.scout_api
@@ -133,94 +133,6 @@ def scout_finished_cases(
         context.invoke(hk_alignment_files, bundle=bundle, yes=yes, dry_run=dry_run)
 
 
-def has_protected_tags(file: hk_models.File, protected_tags_lists: List[List[str]]) -> bool:
-    """Check if a file has any protected tags"""
-
-    LOG.info(f"File {file.full_path} has the tags {file.tags}")
-    version_file_tags: List[str] = HousekeeperAPI.get_tag_names_from_file(file)
-
-    _has_protected_tags: bool = False
-    for protected_tags in protected_tags_lists:
-
-        protected_tags_on_file = set(version_file_tags).intersection(protected_tags)
-        if protected_tags_on_file:
-            if protected_tags_on_file == set(protected_tags):
-                LOG.debug(
-                    "File %s has the protected tag(s) %s, skipping.",
-                    file.full_path,
-                    protected_tags_on_file,
-                )
-                _has_protected_tags = True
-                break
-
-    if not _has_protected_tags:
-        LOG.info("File %s has no protected tags.", file.full_path)
-    return _has_protected_tags
-
-
-def get_bundle_files(
-        housekeeper_api: HousekeeperAPI, bundle_name: str, started_at: datetime, pipeline: Pipeline
-) -> List[hk_models.File]:
-    """Get any bundle files for a specific version"""
-    hk_bundle_version: Optional[hk_models.Version] = housekeeper_api.version(
-        bundle=bundle_name, date=started_at
-    )
-    if not hk_bundle_version:
-        LOG.warning(
-            f"Version not found for "
-            f"bundle:{bundle_name}; "
-            f"pipeline: {pipeline}; "
-            f"date {started_at}"
-        )
-        return []
-
-    LOG.info(
-        f"Version found for "
-        f"bundle:{bundle_name}; "
-        f"pipeline: {pipeline}; "
-        f"date {started_at}"
-    )
-    return housekeeper_api.get_files(bundle=bundle_name, version=hk_bundle_version.id).all()
-
-
-def get_unprotected_existing_bundle_files(status_db: Store, housekeeper_api: HousekeeperAPI, before: datetime) -> \
-Iterator[hk_models.File]:
-    """Returns all existing bundle files from analyses started before 'before' that hos no protected tags"""
-
-    pipeline: Pipeline
-    for pipeline in Pipeline:
-
-        protected_tags_lists = WORKFLOW_PROTECTED_TAGS.get(pipeline)
-        if not protected_tags_lists:
-            LOG.debug("No protected tags defined for %s, skipping", pipeline)
-            continue
-
-        analyses: Query = status_db.get_analyses_before_date(pipeline=pipeline, before=before)
-
-        analysis: models.Analysis
-        for analysis in analyses:
-            LOG.info("Cleaning analysis %s", analysis)
-            version_files: List[hk_models.File] = get_bundle_files(
-                housekeeper_api=housekeeper_api,
-                bundle_name=analysis.family.internal_id,
-                started_at=analysis.started_at,
-                pipeline=pipeline,
-            )
-
-            version_file: hk_models.File
-            for version_file in version_files:
-
-                if has_protected_tags(version_file, protected_tags_lists=protected_tags_lists):
-                    continue
-
-                file_path: Path = Path(version_file.full_path)
-                if not file_path.exists():
-                    LOG.info("File %s not on disk.", file_path)
-                    continue
-                LOG.info("File %s found on disk.", file_path)
-                yield version_file
-
-
 @clean.command("hk-case-bundle-files")
 @click.option(
     "--days-old",
@@ -233,12 +145,13 @@ Iterator[hk_models.File]:
 def hk_case_bundle_files(context: CGConfig, days_old: int, dry_run: bool = False):
     """Clean up all files for all pipelines but those whitelisted"""
     housekeeper_api: HousekeeperAPI = context.obj.housekeeper_api
+    clean_api: CleanAPI = CleanAPI(status_db=context.obj.status_db, housekeeper_api=housekeeper_api)
 
     size_cleaned: int = 0
     version_file: hk_models.File
-    for version_file in get_unprotected_existing_bundle_files(status_db=context.obj.status_db,
-                                                              housekeeper_api=housekeeper_api,
-                                                              before=get_date_days_ago(days_ago=days_old)):
+    for version_file in clean_api.get_unprotected_existing_bundle_files(
+        before=get_date_days_ago(days_ago=days_old)
+    ):
 
         file_path: Path = Path(version_file.full_path)
         file_size = file_path.stat().st_size
@@ -263,12 +176,12 @@ def hk_case_bundle_files(context: CGConfig, days_old: int, dry_run: bool = False
 @DRY_RUN
 @click.pass_obj
 def hk_bundle_files(
-        context: CGConfig,
-        case_id: Optional[str],
-        tags: list,
-        days_old: Optional[int],
-        pipeline: Optional[Pipeline],
-        dry_run: bool,
+    context: CGConfig,
+    case_id: Optional[str],
+    tags: list,
+    days_old: Optional[int],
+    pipeline: Optional[Pipeline],
+    dry_run: bool,
 ):
     """Remove files found in housekeeper bundles"""
