+++ tests/cli/clean/test_hk_case_bundle_files.py
@@ -1,15 +1,16 @@
 import datetime as dt
 import logging
 
-from cg.cli.clean import hk_case_bundle_files, WORKFLOW_PROTECTED_TAGS
+from cg.cli.clean import hk_case_bundle_files
+from cg.constants.tags import WORKFLOW_PROTECTED_TAGS
 from cg.models.cg_config import CGConfig
 from cg.store import Store
 from cg.store import models
 from cgmodels.cg.constants import Pipeline
 from click.testing import CliRunner
-from tests.store_helpers import StoreHelpers
-
 from cg.cli.clean import get_date_days_ago
+
+from tests.store_helpers import StoreHelpers
 
 
 def test_date_days_ago_zero_days():
